refactor - move serializers into separate files
