+++ papapay/user/serializers/user_profile_serializer.py
@@ -1,0 +1,15 @@
+from django.contrib.auth import get_user_model
+
+from rest_framework import serializers
+
+User = get_user_model()
+
+
+class UserProfileSerializer(serializers.ModelSerializer):
+    email = serializers.CharField()
+    first_name = serializers.CharField()
+    last_name = serializers.CharField()
+
+    class Meta:
+        model = User
+        fields = ('email', 'first_name', 'last_name')
